code refining

+++ src/main/java/com/rivigo/repository/RequestRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface RequestRepository extends JpaRepository<Request,Long> {
+public interface RequestRepository extends JpaRepository<Request,Integer> {
     List<Request>findAll();
 }
